Refactor message rendering system and fix tool execution display
- Replace MessageCard and StreamingMessageItem with UnifiedMessageCard
- Implement MessageRenderer with simplified component selection logic
- Fix tool execution results not displaying properly
- Migrate ToolApprovalCard to use Ant Design components
- Add comprehensive documentation for refactoring process
- Improve message type detection and routing
