+++ memory-bank/TOOL_APPROVAL_ANTD_MIGRATION.md
@@ -1,0 +1,145 @@
+# ToolApprovalCard AntD Migration Summary
+
+## 完成日期
+2025年5月26日
+
+## 迁移概述
+成功将 ToolApprovalCard 组件从自定义 CSS 样式迁移到 AntD 原生组件，实现了设计统一和维护性提升。
+
+## 迁移的组件
+
+### 1. ToolApprovalCard (`src/components/ChatView/ToolApp/ToolApprovalCard/index.tsx`)
+
+#### 迁移前问题
+- 使用大量自定义 CSS (~200行)
+- 硬编码的颜色值和间距
+- 手动实现的 dark mode 支持
+- 响应式设计需要手动维护
+- 与项目其他组件风格不一致
+
+#### 迁移后改进
+- **完全使用 AntD 组件**：
+  - `Card` 替代自定义容器
+  - `Space` 进行布局管理
+  - `Typography.Title` 和 `Typography.Text` 显示文本
+  - `Tag` 显示状态标签
+  - `Button` 处理交互
+  - `Alert` 和 `Descriptions` 显示参数
+
+- **设计令牌使用**：
+  - 使用 `theme.useToken()` 获取所有样式值
+  - 自动支持 light/dark 主题
+  - 响应式布局自动处理
+
+- **组件映射**：
+  ```tsx
+  // 旧结构 → 新结构
+  .tool-approval-card → Card (with hoverable)
+  .tool-header → Space (horizontal)
+  .tool-icon → emoji in div
+  .tool-info → Space (vertical)
+  .tool-name → Typography.Title (level 5)
+  .tool-type → Tag (color="blue")
+  .approval-required → Tag (color="error")
+  .tool-content → Alert/Descriptions
+  .tool-parameter.command → Alert with Typography.Text (code)
+  .tool-actions → Space (justify="flex-end")
+  .approve-button → Button (type="primary", icon)
+  .reject-button → Button (icon)
+  ```
+
+#### 参数展示优化
+- **命令参数**：使用 `Alert` 组件 + `Typography.Text code` 样式
+- **其他参数**：使用 `Descriptions` 组件进行结构化展示
+- **样式一致性**：所有样式都使用 design tokens
+
+### 2. ToolCallsSection (`src/components/ChatView/Message/shared/ToolCallsSection.tsx`)
+
+#### 迁移内容
+- 移除了有问题的 `Collapse` 结构
+- 使用简洁的直接展示方式（按用户要求不要折叠）
+- 添加了描述性标题文本
+- 使用 `Space` 组件管理布局
+- 完全使用 design tokens
+
+#### 新结构
+```tsx
+<div style={{ marginTop: token.marginMD }}>
+  <Typography.Text type="secondary">
+    检测到 {toolCalls.length} 个工具调用
+  </Typography.Text>
+  <Space direction="vertical" size="middle">
+    {/* ToolApprovalCard 列表 */}
+  </Space>
+</div>
+```
+
+## 技术改进
+
+### 1. 代码质量
+- **代码量减少**：删除了 ~200 行自定义 CSS
+- **类型安全**：完全使用 TypeScript + AntD 类型
+- **一致性**：与项目其他组件风格统一
+
+### 2. 主题支持
+- **自动主题切换**：无需手动处理 dark mode
+- **设计令牌**：所有样式值来自 AntD theme
+- **响应式**：自动适配不同屏幕尺寸
+
+### 3. 维护性
+- **组件复用**：使用 AntD 成熟组件
+- **更新容易**：跟随 AntD 版本自动更新
+- **调试友好**：使用标准组件更容易调试
+
+## 文件变更
+
+### 删除的文件
+- `src/components/ChatView/ToolApp/ToolApprovalCard/styles.css`
+
+### 修改的文件
+- `src/components/ChatView/ToolApp/ToolApprovalCard/index.tsx`
+- `src/components/ChatView/Message/shared/ToolCallsSection.tsx`
+
+## 测试结果
+- ✅ 开发服务器启动成功
+- ✅ 组件编译无错误
+- ✅ TypeScript 类型检查通过
+- ✅ 功能保持完整（approve/reject 逻辑）
+
+## 用户体验改进
+
+### 视觉改进
+- **更好的视觉层次**：使用 AntD 的设计语言
+- **一致的间距**：所有间距使用 design tokens
+- **适当的颜色**：状态颜色更加语义化
+- **更好的 hover 效果**：Card 组件的内置 hover 状态
+
+### 交互改进
+- **更清晰的状态指示**：使用颜色化的 Tag 组件
+- **更好的按钮样式**：主操作使用 primary 按钮
+- **图标支持**：按钮添加了语义化图标
+- **更好的参数展示**：command 参数使用专门的展示方式
+
+## 后续计划
+项目中还发现其他使用自定义 CSS 的组件：
+- `src/components/ChatView/ChatView.css`
+- `src/components/Shared/SearchWindow/styles.css`
+- `src/components/Sidebar/ChatItem/styles.css`
+
+可以考虑将这些组件也迁移到 AntD 以保持设计一致性。
+
+## 关键学习点
+
+### AntD 最佳实践
+1. **始终使用 design tokens**：通过 `theme.useToken()` 获取样式值
+2. **组件组合**：使用 `Space` + 其他组件构建复杂布局
+3. **语义化标签**：使用 `Tag` 的 color 属性表示状态
+4. **适当的组件选择**：`Alert` 用于重要信息，`Descriptions` 用于结构化数据
+
+### 迁移策略
+1. **逐个组件迁移**：避免大规模重构的风险
+2. **保持接口一致**：确保组件 props 接口不变
+3. **功能验证**：确保所有原有功能正常工作
+4. **渐进式改进**：在迁移过程中适当优化用户体验
+
+这次迁移成功展示了如何在保持功能完整性的同时，显著提升代码质量和设计一致性。
