--- src/components/ChatView/ToolApp/ToolApprovalCard/styles.css
@@ -1,202 +1,0 @@
-.tool-approval-card {
-  border-radius: 8px;
-  border: 1px solid #e6e6e6;
-  padding: 16px;
-  margin-bottom: 12px;
-  background-color: #fff;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
-  transition: all 0.3s ease;
-}
-
-.tool-approval-card:hover {
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-}
-
-.tool-approval-card.requires-approval {
-  border-left: 4px solid #f5222d;
-}
-
-.tool-approval-card.safe-tool {
-  border-left: 4px solid #52c41a;
-}
-
-.tool-header {
-  display: flex;
-  align-items: center;
-  margin-bottom: 12px;
-}
-
-.tool-icon {
-  font-size: 24px;
-  margin-right: 12px;
-}
-
-.tool-info {
-  flex: 1;
-}
-
-.tool-name {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.tool-type {
-  font-size: 12px;
-  color: #8c8c8c;
-  display: inline-block;
-  margin-right: 8px;
-}
-
-.approval-required {
-  background-color: #fff1f0;
-  color: #f5222d;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 500;
-}
-
-.executing {
-  background-color: #e6f7ff;
-  color: #1677ff;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 500;
-}
-
-.rejected {
-  background-color: #f6f6f6;
-  color: #8c8c8c;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 500;
-}
-
-.executed {
-  background-color: #f6ffed;
-  color: #52c41a;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-size: 12px;
-  font-weight: 500;
-}
-
-.tool-content {
-  margin-bottom: 16px;
-}
-
-.tool-parameter {
-  margin-bottom: 8px;
-}
-
-.tool-parameter.command {
-  background-color: #f5f5f5;
-  padding: 8px;
-  border-radius: 4px;
-  overflow-x: auto;
-}
-
-.param-label {
-  font-weight: 500;
-  margin-right: 8px;
-}
-
-.command-code {
-  display: block;
-  background-color: #2c2c2c;
-  color: #e6e6e6;
-  padding: 8px;
-  border-radius: 4px;
-  margin-top: 4px;
-  font-family: 'Courier New', monospace;
-  white-space: pre-wrap;
-  word-break: break-all;
-}
-
-.tool-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 8px;
-}
-
-.approve-button, 
-.reject-button {
-  padding: 6px 12px;
-  border-radius: 4px;
-  cursor: pointer;
-  font-weight: 500;
-  border: none;
-  transition: all 0.2s ease;
-}
-
-.approve-button {
-  background-color: #1677ff;
-  color: white;
-}
-
-.approve-button:hover:not(:disabled) {
-  background-color: #4096ff;
-}
-
-.approve-button:disabled {
-  background-color: #d9d9d9;
-  color: #8c8c8c;
-  cursor: not-allowed;
-}
-
-.reject-button {
-  background-color: #f0f0f0;
-  color: #595959;
-}
-
-.reject-button:hover:not(:disabled) {
-  background-color: #e0e0e0;
-}
-
-.reject-button:disabled {
-  background-color: #f5f5f5;
-  color: #bfbfbf;
-  cursor: not-allowed;
-}
-
-/* Dark mode support */
-[data-theme="dark"] .tool-approval-card {
-  background-color: #1f1f1f;
-  border-color: #303030;
-}
-
-[data-theme="dark"] .tool-parameter.command {
-  background-color: #141414;
-}
-
-[data-theme="dark"] .command-code {
-  background-color: #000;
-  color: #d9d9d9;
-}
-
-[data-theme="dark"] .reject-button {
-  background-color: #2a2a2a;
-  color: #d9d9d9;
-}
-
-[data-theme="dark"] .reject-button:hover {
-  background-color: #3a3a3a;
-}
-
-/* 响应式调整 */
-@media (max-width: 600px) {
-  .tool-actions {
-    flex-direction: column;
-  }
-  
-  .tool-parameter {
-    flex-direction: column;
-  }
-  
-  .param-label {
-    margin-bottom: 4px;
-  }
-}
